# A better split-2D(SliceChannel) op forward kernel for CPU