@vandanavk 2019-02-05
@mxnet-label-bot add [pr-awaiting-review, Operator]

@huangzhiyuan Can this solution be extended to support variable length splits as well?